refactor : 주문 모듈 패키지 구조 변경
