--- src/main/java/com/pintoss/auth/module/order/execution/OrderFactory.java
@@ -1,11 +1,0 @@
-package com.pintoss.auth.module.order.execution;
-
-import com.pintoss.auth.module.order.execution.domain.Order;
-
-public class OrderFactory {
-
-    public Order create(){
-        return null;
-    }
-
-}
