+++ src/main/java/com/pintoss/auth/module/order/external/OrderRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.pintoss.auth.module.order.external;
 
-import com.pintoss.auth.module.order.execution.OrderRepository;
-import com.pintoss.auth.module.order.execution.domain.Order;
+import com.pintoss.auth.module.order.usecase.service.OrderRepository;
+import com.pintoss.auth.module.order.model.Order;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
