+++ src/main/java/com/pintoss/auth/module/order/usecase/service/OrderFactory.java
@@ -1,0 +1,11 @@
+package com.pintoss.auth.module.order.usecase.service;
+
+import com.pintoss.auth.module.order.model.Order;
+
+public class OrderFactory {
+
+    public Order create(){
+        return null;
+    }
+
+}
