+++ src/main/java/com/pintoss/auth/module/order/usecase/service/OrderReader.java
@@ -1,6 +1,6 @@
-package com.pintoss.auth.module.order.execution;
+package com.pintoss.auth.module.order.usecase.service;
 
-import com.pintoss.auth.module.order.execution.domain.Order;
+import com.pintoss.auth.module.order.model.Order;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
