+++ src/main/java/com/pintoss/auth/module/order/usecase/service/OrderRepository.java
@@ -1,6 +1,6 @@
-package com.pintoss.auth.module.order.execution;
+package com.pintoss.auth.module.order.usecase.service;
 
-import com.pintoss.auth.module.order.execution.domain.Order;
+import com.pintoss.auth.module.order.model.Order;
 import java.util.Optional;
 
 public interface OrderRepository {
